+++ lib/compose.mjs
@@ -1,4 +1,4 @@
-import curry from './f.curry.mjs'
+import curry from './curry.mjs'
 
 const compose = (...args) =>
   (...rest) =>
