+++ lib/compose.test.js
@@ -1,6 +1,6 @@
 const load = require('@std/esm')(module)
 const test = require('ava')
-const {compose} = load('../lib/f.mjs')
+const compose = load('./compose.mjs').default
 
 const greet = name => `Hello ${name}.`
 const exclaim = message => message.toUpperCase()
