+++ lib/curry.test.js
@@ -1,6 +1,6 @@
 const load = require('@std/esm')(module)
 const test = require('ava')
-const {curry} = load('../lib/f.mjs')
+const curry = load('./curry.mjs').default
 
 const add = (a, b) => a + b
