+++ lib/flip.test.js
@@ -1,6 +1,6 @@
 const load = require('@std/esm')(module)
 const test = require('ava')
-const {flip} = load('../lib/f.mjs')
+const flip = load('./flip.mjs').default
 
 const add = (a, b) => a + b
