+++ lib/merge.mjs
