+++ lib/merge.test.js
@@ -1,6 +1,6 @@
 const load = require('@std/esm')(module)
 const test = require('ava')
-const {merge} = load('../lib/f.mjs')
+const merge = load('./merge.mjs').default
 
 test('curry', t => {
   let a = {
