+++ lib/pipe.test.js
@@ -1,6 +1,6 @@
 const load = require('@std/esm')(module)
 const test = require('ava')
-const {pipe} = load('../lib/f.mjs')
+const pipe = load('./pipe.mjs').default
 
 const greet = name => `Hello ${name}.`
 const exclaim = message => message.toUpperCase()
